# src/main/resources/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Movie Ticket Booking API
  description: 'Browse screenings, reserve seats, complete bookings, cancel bookings, and administer the
    local catalog. This document is the source of truth: request/response models and the controller interfaces
    are generated from it at build time.'
  version: v1
servers:
- url: http://localhost:8080
  description: Local development
...
paths:
...
  /admin/api/v1/refund-policies/{id}:
    get:
      tags:
      - Admin pricing
      operationId: getRefundPolicy
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundPolicyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    put:
      tags:
      - Admin pricing
      operationId: updateRefundPolicy
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundPolicyRequest'
            example:
              name: Standard refund policy
              rules:
              - cutoffMinutes: 1440
                refundPercentage: 100.0
              - cutoffMinutes: 120
                refundPercentage: 50.0
              - cutoffMinutes: 0
                refundPercentage: 0.0
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundPolicyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin pricing
      operationId: deactivateRefundPolicy
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /admin/api/v1/pricing-plans/{id}:
    get:
      tags:
      - Admin pricing
      operationId: getPricingPlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingPlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    put:
      tags:
      - Admin pricing
      operationId: updatePricingPlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PricingPlanRequest'
            example:
              weekendAdjustment: 50.0
              regularPrice: 250.0
              premiumPrice: 400.0
              name: Standard INR
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingPlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin pricing
      operationId: deactivatePricingPlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/discount-codes/{id}:
    get:
      tags:
      - Admin pricing
      operationId: getDiscountCode
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscountCodeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
      - basicAuth: []
    put:
      tags:
      - Admin pricing
      operationId: updateDiscountCode
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountCodeRequest'
            example:
              validUntil: '2030-12-31T23:59:59Z'
              code: WELCOME10
              value: 10.0
              type: PERCENTAGE
              validFrom: '2030-01-01T00:00:00Z'
              globalUsageLimit: 1000
              perCustomerUsageLimit: 1
              minimumSpend: 200.0
              maximumDiscount: 100.0
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscountCodeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
    delete:
      tags:
      - Admin pricing
      operationId: deactivateDiscountCode
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: No Content
          content:
            application/json: {}
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/refund-policies:
    get:
      tags:
      - Admin pricing
      operationId: listRefundPolicies
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseRefundPolicyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin pricing
      operationId: createRefundPolicy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundPolicyRequest'
            example:
              name: Standard refund policy
              rules:
              - cutoffMinutes: 1440
                refundPercentage: 100.0
              - cutoffMinutes: 120
                refundPercentage: 50.0
              - cutoffMinutes: 0
                refundPercentage: 0.0
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundPolicyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
  /admin/api/v1/pricing-plans:
    get:
      tags:
      - Admin pricing
      operationId: listPricingPlans
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponsePricingPlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin pricing
      operationId: createPricingPlan
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PricingPlanRequest'
            example:
              weekendAdjustment: 50.0
              regularPrice: 250.0
              premiumPrice: 400.0
              name: Standard INR
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingPlanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
  /admin/api/v1/discount-codes:
    get:
      tags:
      - Admin pricing
      operationId: listDiscountCodes
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseDiscountCodeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
      security:
      - basicAuth: []
    post:
      tags:
      - Admin pricing
      operationId: createDiscountCode
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountCodeRequest'
            example:
              validUntil: '2030-12-31T23:59:59Z'
              code: WELCOME10
              value: 10.0
              type: PERCENTAGE
              validFrom: '2030-01-01T00:00:00Z'
              globalUsageLimit: 1000
              perCustomerUsageLimit: 1
              minimumSpend: 200.0
              maximumDiscount: 100.0
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscountCodeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/BusinessViolation'
      security:
      - basicAuth: []
...
components:
  schemas:
...
    RefundPolicyRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
        rules:
          type: array
          items:
            $ref: '#/components/schemas/RefundRuleRequest'
          maxItems: 20
          minItems: 1
      required:
      - name
      - rules
    RefundRuleRequest:
      type: object
      properties:
        cutoffMinutes:
          type: integer
          format: int64
          minimum: 0
        refundPercentage:
          type: number
          maximum: 100.0
          minimum: 0.0
          x-field-extra-annotation: '@jakarta.validation.constraints.Digits(integer = 3, fraction = 2)'
      required:
      - cutoffMinutes
      - refundPercentage
    RefundPolicyResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        rules:
          type: array
          items:
            $ref: '#/components/schemas/RefundRuleResponse'
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    RefundRuleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        cutoffMinutes:
          type: integer
          format: int64
        refundPercentage:
          type: number
    PricingPlanRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
        regularPrice:
          type: number
          minimum: 0.01
          x-field-extra-annotation: '@jakarta.validation.constraints.Digits(integer = 10, fraction = 2)'
        premiumPrice:
          type: number
          minimum: 0.01
          x-field-extra-annotation: '@jakarta.validation.constraints.Digits(integer = 10, fraction = 2)'
        weekendAdjustment:
          type: number
          minimum: 0.0
          x-field-extra-annotation: '@jakarta.validation.constraints.Digits(integer = 10, fraction = 2)'
      required:
      - name
      - premiumPrice
      - regularPrice
      - weekendAdjustment
    PricingPlanResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        regularPrice:
          type: number
        premiumPrice:
          type: number
        weekendAdjustment:
          type: number
        currency:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
    DiscountCodeRequest:
      type: object
      properties:
        code:
          type: string
          maxLength: 40
          minLength: 1
        type:
          $ref: '#/components/schemas/DiscountType'
        value:
          type: number
          minimum: 0.01
          x-field-extra-annotation: '@jakarta.validation.constraints.Digits(integer = 10, fraction = 2)'
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        minimumSpend:
          type: number
          minimum: 0.0
          x-field-extra-annotation: '@jakarta.validation.constraints.Digits(integer = 10, fraction = 2)'
        maximumDiscount:
          type: number
          minimum: 0.01
          x-field-extra-annotation: '@jakarta.validation.constraints.Digits(integer = 10, fraction = 2)'
        globalUsageLimit:
          type: integer
          format: int32
          minimum: 1
        perCustomerUsageLimit:
          type: integer
          format: int32
          minimum: 1
      required:
      - code
      - minimumSpend
      - type
      - validFrom
      - validUntil
      - value
    DiscountCodeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        type:
          $ref: '#/components/schemas/DiscountType'
        value:
          type: number
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: string
          format: date-time
        minimumSpend:
          type: number
        maximumDiscount:
          type: number
        globalUsageLimit:
          type: integer
          format: int32
        perCustomerUsageLimit:
          type: integer
          format: int32
        currency:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
...
    PageResponseRefundPolicyResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RefundPolicyResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageResponsePricingPlanResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PricingPlanResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    PageResponseDiscountCodeResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DiscountCodeResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    FieldViolation:
      type: object
      properties:
        field:
          type: string
          example: screeningSeatIds[1]
        code:
          type: string
          example: NotNull
        message:
          type: string
          example: The field is required.
      required:
      - code
      - field
      - message
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: /problems/seat-unavailable
        title:
          type: string
          example: Seat unavailable
        status:
          type: integer
          format: int32
          example: 409
        detail:
          type: string
          example: One or more requested seats are no longer available.
        instance:
          type: string
          format: uri
          example: /api/v1/seat-reservations
        code:
          type: string
          example: SEAT_UNAVAILABLE
        requestId:
          type: string
          format: uuid
          example: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
      required:
      - code
      - detail
      - instance
      - requestId
      - status
      - title
      - type
...
    DiscountType:
      type: string
      enum:
      - FIXED
      - PERCENTAGE
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Invalid request
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/invalid-request
            status: 400
            instance: /api/v1/seat-reservations
            code: INVALID_REQUEST
            detail: The request could not be processed.
    Unauthenticated:
      description: Valid HTTP Basic credentials are required.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Authentication required
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/authentication-required
            status: 401
            instance: /api/v1/seat-reservations
            code: AUTHENTICATION_REQUIRED
            detail: The request could not be processed.
    Forbidden:
      description: The authenticated account lacks the required role.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Access denied
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/access-denied
            status: 403
            instance: /api/v1/seat-reservations
            code: ACCESS_DENIED
            detail: The request could not be processed.
    NotFound:
      description: The resource does not exist or is concealed from this account.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Resource not found
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/resource-not-found
            status: 404
            instance: /api/v1/seat-reservations
            code: RESOURCE_NOT_FOUND
            detail: The request could not be processed.
    Conflict:
      description: The request conflicts with the current resource state.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Request conflict
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/request-conflict
            status: 409
            instance: /api/v1/seat-reservations
            code: REQUEST_CONFLICT
            detail: The request could not be processed.
    BusinessViolation:
      description: A business rule prevents the operation.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Business rule violation
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/business-rule-violation
            status: 422
            instance: /api/v1/seat-reservations
            code: BUSINESS_RULE_VIOLATION
            detail: The request could not be processed.
    InternalError:
      description: An unexpected failure occurred.
      headers:
        X-Request-Id:
          description: Server-generated request correlation identifier
          schema:
            type: string
            format: uuid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            title: Internal server error
            requestId: 8f3d8cbe-0e42-4d9b-9d50-96199b173ba7
            type: /problems/internal-error
            status: 500
            instance: /api/v1/seat-reservations
            code: INTERNAL_ERROR
            detail: The request could not be processed.
  parameters:
...
    Page:
      name: page
      in: query
      description: Zero-based page index
      schema:
        type: integer
        format: int32
        default: 0
    PageSize:
      name: size
      in: query
      description: Number of records from 1 to 100
      schema:
        type: integer
        format: int32
        default: 20
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic